Rename project to D-Cell
The project was tentatively titled `tiny-expr`, but it's neither tiny
nor a particularly unique name (there are several projects with the same
name in C).

Instead, the new name D-Cell is taken, since this takes some level of
influence from the Google CEL project, which is a DSL for evaluating
conditions.