Add package for integer-comparisons
The D-Cell language doesn't distinguish between signed and unsigned
integers for comparison so that the code will "do the right thing".
To make this possible, we need to implement some logic that
performs signed and unsigned comparisons in a rational way.

This adds a package that does exactly that; when faced with a signed
and unsigned comparison, it will determine whether the value exceeds
the range and returns the appropriate result, so that no unexpected
integer promotions occur.