Add integer conversion utilities
D-Cell does not make any distinction between the size of integers,
but it's desired to still be capable of narrowing the values back down
to different integer sizes. This adds a package that makes it easier
to determine unexpected overflows and underflows when converting between
signed and unsigned integers, or when shrinking the size of an integer.