Add package for comparing `reflect.Value`s
The previous change introduced a new package for converting
`reflect.Value` types to higher level types. This was a necessary step
to enable the comparison of `reflect.Value` types, since the comparisons
are done by semantic values rather than representational equality.

This will enable equality and inequality operators in the grammar to be
implemented later.